updating e2fsprogs

+++ src/e2fsprogs/lib/e2p/hashstr.c
@@ -1,13 +1,15 @@
 /*
  * feature.c --- convert between features and strings
- * 
+ *
  * Copyright (C) 1999  Theodore Ts'o <tytso@mit.edu>
- * 
- * This file can be redistributed under the terms of the GNU Library General
- * Public License
- * 
+ *
+ * %Begin-Header%
+ * This file may be redistributed under the terms of the GNU Library
+ * General Public License, version 2.
+ * %End-Header%
  */
 
+#include "config.h"
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
@@ -48,7 +50,7 @@ int e2p_string2hash(char *string)
 {
 	struct hash	*p;
 	char		*eptr;
-	long		num;
+	int		num;
 
 	for (p = hash_list; p->string; p++) {
 		if (!strcasecmp(string, p->string)) {
